third_party/WebKit/LayoutTests/external/wpt/web-animations/timing-model/time-transformations/transformed-progress-expected.txt

This is a testharness.js-based test.
PASS Transformed progress for step-start function
PASS Transformed progress for steps(1, start) function
PASS Transformed progress for steps(2, start) function
PASS Transformed progress for step-end function
PASS Transformed progress for steps(1) function
PASS Transformed progress for steps(1, end) function
PASS Transformed progress for steps(2, end) function
PASS Transformed progress for frames function
PASS Transformed progress for linear function
FAIL Transformed progress for ease function assert_approx_equals: The progress should be approximately 0.40851059137130497 at 250ms expected 0.40851059137130497 +/- 0.01 but got 0.41869212962962965
PASS Transformed progress for ease-in function
PASS Transformed progress for ease-in-out function
PASS Transformed progress for ease-out function
PASS Transformed progress for easing function which produces values greater than 1
PASS Transformed progress for easing function which produces values less than 1
FAIL Test bounds point of step-start easing assert_equals: Progress at 0ms expected 0 but got 0.5
FAIL Test bounds point of step-start easing with reverse direction assert_equals: Progress at 2000ms expected 0 but got 0.5
PASS Test bounds point of step-start easing with iterationStart not at a transition point
FAIL Test bounds point of step-start easing with iterationStart and delay assert_equals: Progress at 0ms expected 0.5 but got 1
FAIL Test bounds point of step-start easing with iterationStart and reverse direction assert_equals: Progress at 2000ms expected 0.5 but got 1
FAIL Test bounds point of step(4, start) easing with iterationStart 0.75 and delay assert_equals: Progress at 0ms expected 0.75 but got 1
FAIL Test bounds point of step-start easing with alternate direction assert_equals: Progress at 3000ms expected 0.5 but got 1
FAIL Test bounds point of step-start easing with alternate-reverse direction assert_equals: Progress at 3000ms expected 0.5 but got 1
PASS Test bounds point of step-end easing
FAIL Test bounds point of step-end easing with iterationStart and delay assert_equals: Progress at 0ms expected 0 but got 0.5
PASS Test bounds point of step-end easing with iterationStart not at a transition point
PASS Test bounds point of frames easing
PASS Test bounds point of frames easing with iterationStart and delay
Harness: the test ran to completion.
